pyproject.toml
[tool.poetry]
name = "pruningg"
version = "0.1.0"
description = ""
authors = ["ciaran-regan-ie <ciaranregan123@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
transformers = "^4.44.2"
trl = "^0.9.6"
torch = "^2.4.0"
python-dotenv = "^1.0.1"
peft = "^0.12.0"
datasets = "^2.21.0"
wandb = "^0.18.3"
evaluate = "^0.4.3"
scikit-learn = "^1.5.2"
deepeval = "^1.3.5"
matplotlib = "^3.9.2"
seaborn = "^0.13.2"
unsloth = "^2024.10.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
